+++ pyutils/__init__.py
@@ -1,5 +1,5 @@
 from .color import BaseColor, Color, ResetColor, RGBColor, BackgroundColor, Colors, TraceBackColor, ColorPalette
-from .configFile import ConfigFile, RaiseType, Profile
+from .configFile import ConfigFile, RaiseType, Profile, Default, ConfigFormat, Option, Options
 from .feedback import FeedBack, Loading, eprint
 from .logger import Logger, LoggerType, LogConfig, FatalFailure
 from .progress import progress, prange