+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - ```Color```: Now themes have a `.reset` attribute by default
 - ```ConfigFile```: Now have config options: Can get different variant of a config format
 - ```ConfigFile```: Now support default values
+- ```progress```: Fixed the progress that did not go all the way to the max when near 100%
 ## 0.0.3
 - ```progress bar```: Fixed the auto-sizing (Now it should be accurate) + Remove trailing characters
 ## 0.0.2